refactor(network): move event handling to its own module
