+++ sn_networking/src/lib.rs
@@ -15,7 +15,6 @@ mod cmd;
 mod driver;
 mod error;
 mod event;
-mod get_record_handler;
 mod log_markers;
 #[cfg(feature = "open-metrics")]
 mod metrics;
